% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_uhi_hist_data}
\alias{add_uhi_hist_data}
\title{Add UHI data from RasterLayer stack to sf data frame}
\usage{
add_uhi_hist_data(uhi_stack_list, sf_data)
}
\arguments{
\item{uhi_stack_list}{List, nested on time of year, time of day}

\item{sf_data}{sf points, data frame with Berlin tree locations and meta data}
}
\value{
A nested list containing RasterLayers with the following levels:
\enumerate{
\item{Time of year}{Summer / Winter}
\item{Time of day}{Night / Day}
}
Containing sf point objects (multiple copies of full data set - careful! \strong{needs improvement!})
}
\description{
Add UHI data from RasterLayer stack to sf data frame
}
