% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_models}
\alias{apply_models}
\title{Apply models (lme4)}
\usage{
apply_models(full_df, extract_uhi, model_list, n_top_species = 6, ...)
}
\arguments{
\item{full_df}{cleaned data set}

\item{extract_uhi}{extracted UHI data}

\item{model_list}{list of models to apply}

\item{...}{an optional unquoted expression passed to \code{dplyr::filter}}
}
\value{
a list containing lme4 model outputs
}
\description{
Apply models (lme4)
}
