% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bind_rows_sf}
\alias{bind_rows_sf}
\title{Bind rows of sf tibbles to single data set}
\usage{
bind_rows_sf(sf_list)
}
\arguments{
\item{sf_list}{list containing simple feature tibbles}
}
\value{
single sf-tibble
}
\description{
Bind rows of sf tibbles to single data set
}
