% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_uhi_stats}
\alias{calc_uhi_stats}
\title{Calculate descriptive statistics for a UHI stacks (in lists)}
\usage{
calc_uhi_stats(uhi_stack_list)
}
\arguments{
\item{uhi_stack_list}{A two-level list (time of year, time of day) of UHI data}
}
\value{
A list containing Median, Mean and StDev of UHI intensities by RasterLayer (years)
}
\description{
Calculate descriptive statistics for a UHI stacks (in lists)
}
