% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_genus}
\alias{clean_genus}
\title{Clean Feature Meta Data}
\usage{
clean_genus(sf_data)
}
\arguments{
\item{sf_data}{sf-tibble of Berlin City trees}
}
\value{
sf tibble with cleaned genus names
}
\description{
Formats upper/lower case and removes "" as genera; still need to check if other
problematic values exist or can be recovered from other meta data.
}
