% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crop_data_with_bbox}
\alias{crop_data_with_bbox}
\title{Crop Spatial data with bbox}
\usage{
crop_data_with_bbox(sf_data_list, bbox)
}
\arguments{
\item{sf_data_list}{list containing sf tibbles}

\item{bbox}{sf tibble, polygon used for cropping}
}
\value{
sf tibble, cropped with specified bbox
}
\description{
Crop Spatial data with bbox
}
