% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dens_plot_trees}
\alias{dens_plot_trees}
\title{Density plot overview}
\usage{
dens_plot_trees(sf_data,
  extracted_uhi,
  position_stack = "stack",
  ymin = 0.4,
  ymax=1.4,
  xmin=-5,
  xmax=-0.5)
}
\arguments{
\item{sf_data}{sf, data set containing Berlin trees}

\item{extracted_uhi}{List, contains UHI intensities for Summer/Winter and day/night.}

\item{position_stack}{Character, either "stack" or "dodge".}

\item{ymin}{Numeric, coordinate for inset plot.}

\item{ymax}{Numeric, coordinate for inset plot.}

\item{xmin}{Numeric, coordinate for inset plot.}

\item{xmax}{Numeric, coordinate for inset plot.}
}
\value{
A ggplot2 plot object
}
\description{
This function is hard-codes use of 2007 Summer, Day UHI data.
}
