% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_berlin_baumscheiben}
\alias{download_berlin_baumscheiben}
\title{Download baumscheiben from Berlin daten WFS}
\usage{
download_berlin_baumscheiben()
}
\value{
Baumscheiben JSON
}
\description{
Download baumscheiben from Berlin daten WFS
}
