% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_berlin_trees}
\alias{download_berlin_trees}
\title{Download Trees from daten.berlin.de}
\usage{
download_berlin_trees()
}
\description{
Download Trees from daten.berlin.de
}
\details{
Downloads tree data from WFS Service for all packages containing 'baum',
excluding the Baumscheiben data set.
}
