% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_maybe}
\alias{filter_maybe}
\title{Adjusted filtering with ...}
\usage{
filter_maybe(df, ...)
}
\arguments{
\item{df}{data frame / tibble for filtering}

\item{...}{unquoted expression for filtering}
}
\value{
filtered df
}
\description{
Adjusted filtering with ...
}
