% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_berlin_polygons_as_sf}
\alias{get_berlin_polygons_as_sf}
\title{Access Berlin district polygon from external source}
\usage{
get_berlin_polygons_as_sf()
}
\value{
sf-tibble with polygons of Berlin districts
}
\description{
Access Berlin district polygon from external source
}
