% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_uhi_rasters}
\alias{get_uhi_rasters}
\title{Extract UHI rasters from tifs}
\usage{
get_uhi_rasters(path)
}
\arguments{
\item{path}{Character, path to folder containing UHI zip archives.}
}
\value{
A nested list containing RasterLayers with the following levels:
\enumerate{
\item{Time of year}{Summer / Winter}
\item{Time of day}{Night / Day}
}
For the Years 2003 to 2017.
}
\description{
Extracts summer and night time UHI gridded data from tifs.
Data was accessed via the UHI Explorer App (2020-01-08;
\url{https://yceo.users.earthengine.app/view/uhimap}) for the Berlin Urban Cluster.
}
\references{
Chakraborty, T., and X. Lee.
A Simplified Urban-Extent Algorithm to Characterize Surface Urban Heat Islands
on a Global Scale and Examine Vegetation Control on Their Spatiotemporal Variability’.
International Journal of Applied Earth Observation and Geoinformation 74 (February 2019):
269–80. \url{https://doi.org/10.1016/j.jag.2018.09.015}.
}
