% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_downloaded_data_to_lists}
\alias{load_downloaded_data_to_lists}
\title{Load spatial-features data sets of all Berlin trees}
\usage{
load_downloaded_data_to_lists(download_data)
}
\arguments{
\item{download_data}{Character, file paths for downloaded .geojson files from WFS service}
}
\value{
list of sf-tibbles
}
\description{
Load spatial-features data sets of all Berlin trees
}
