% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_age_table}
\alias{make_age_table}
\title{Generate table of genera age distribution}
\usage{
make_age_table(df, max_age = 120, break_interval = 20)
}
\arguments{
\item{df}{Berlin trees data frame}

\item{max_age}{numeric, last cut}

\item{break_interval}{numeric, size of intervals, note, max_age/break_intervals must be an integer.}
}
\value{

}
\description{
Generate table of genera age distribution
}
