% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_bbox}
\alias{make_bbox}
\title{Define bounding box as area of interest}
\usage{
make_bbox(lat_min, lat_max, lon_min, lon_max, feature_name, crs = 4326)
}
\arguments{
\item{lat_min}{numeric}

\item{lat_max}{numeric}

\item{lon_min}{numeric}

\item{lon_max}{numeric}

\item{feature_name}{character, name of sf feature}

\item{crs}{numeric, defaults to 4326 (lat-lon Mercator, WGS84)}
}
\value{
sf polygon
}
\description{
Define bounding box as area of interest
}
