% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_overview_map}
\alias{make_overview_map}
\title{Create overview-map of all data sets}
\usage{
make_overview_map(sf_data, poly)
}
\arguments{
\item{sf_data}{sf-tibble of Berlin trees}

\item{poly}{sf-tibble, polygons of Berlin districts}
}
\value{
ggplot object
}
\description{
Create overview-map of all data sets
}
