% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_overview_table}
\alias{make_overview_table}
\title{Make overview table}
\usage{
make_overview_table(df)
}
\arguments{
\item{df}{berlin trees df}
}
\value{
data frame ready for kabling
}
\description{
Make overview table
}
