% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_test_data_set}
\alias{make_test_data_set}
\title{Generate data set for stat-models}
\usage{
make_test_data_set(full_df = full_data_set_clean,
  extract_uhi = extract_uhi_values_to_list)
}
\arguments{
\item{full_df}{cleaned data.frame of berlin trees}

\item{extract_uhi}{list of sf data frames with UHI values}
}
\value{
a data frame for use with \code{\link{apply_models}}
}
\description{
Generate data set for stat-models
}
