% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_uhi_plot}
\alias{make_uhi_plot}
\title{Plot UHI with Berlin districts}
\usage{
make_uhi_plot(
  uhi_stacks,
  berlin_poly,
  base_size = 18,
  file,
  height,
  width,
  dpi
)
}
\arguments{
\item{uhi_stacks}{Rasterstack lists for UHI}

\item{berlin_poly}{Berlin District Polygon}

\item{base_size}{Numeric, base char size for ggplot}

\item{file}{Character, output file path}

\item{height}{Numeric, for plot output (cm)}

\item{width}{Numeric, for plot output (cm)}

\item{dpi}{Numeric}
}
\value{

}
\description{
Plot UHI with Berlin districts
}
