% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tree_sums_bar_plot}
\alias{tree_sums_bar_plot}
\title{Generate overview of records (bar plot)}
\usage{
tree_sums_bar_plot(sf_data)
}
\arguments{
\item{sf_data}{sf-tibble of Berlin City trees}
}
\value{
ggplot bar graph
}
\description{
Generate overview of records (bar plot)
}
