% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{apply_models}
\alias{apply_models}
\title{Apply models (lme4)}
\usage{
apply_models(
  df = test_df,
  model_list = model_list,
  n_top_species = 6,
  min_individuals = 150
)
}
\arguments{
\item{df}{data.frame containing extracted uhi values and tree data}

\item{model_list}{list of models to apply}

\item{n_top_species}{numeric, number of top species to consider}

\item{min_individuals}{numeric, min. individuals per species required
for inclusion}
}
\value{
a list containing lme4 model outputs
}
\description{
Apply models (lme4)
}
