% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_ranef_plot}
\alias{make_ranef_plot}
\title{Make Random-effects effect-size plot}
\usage{
make_ranef_plot(
  model_out,
  model_name = "heat_RIspecies_RSspecies_RIprovenance",
  df,
  n_top_species = 6
)
}
\arguments{
\item{model_out}{output list with models}

\item{model_name}{Character, model name}

\item{df}{data.frame with all trees}

\item{n_top_species}{Numeric, max number of species to inspect}
}
\value{
ggplot2 object and plot
}
\description{
Make Random-effects effect-size plot
}
