% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_ranef_plot}
\alias{make_ranef_plot}
\title{Make Random-effects effect-size plot}
\usage{
make_ranef_plot(model_out,
  model_name = "heat_RIspecies_RSspecies_RIprovenance",
  n_top_species = 6, full_df = full_data_set_clean,
  extract_uhi = extract_uhi_values_to_list)
}
\arguments{
\item{model_out}{output list with models}

\item{model_name}{Character, model name}
}
\value{
ggplot2 object and plot
}
\description{
Make Random-effects effect-size plot
}
