% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tree_count_map}
\alias{tree_count_map}
\title{Spatially-binned tree counts}
\usage{
tree_count_map(sf_data, poly)
}
\arguments{
\item{sf_data}{sf-tibble of Berlin City trees}

\item{poly}{sf-tibble of Berlin Districts}
}
\value{
ggplot map
}
\description{
Spatially-binned tree counts
}
