% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tree_sums_bar_plot}
\alias{tree_sums_bar_plot}
\title{Generate overview of records (bar plot)}
\usage{
tree_sums_bar_plot(sf_data, base_size = 18, file, height, width, dpi)
}
\arguments{
\item{sf_data}{sf-tibble of Berlin City trees}

\item{file}{character, file path (use with \code{\link{file_out}}), must include file ending}

\item{height}{numeric, height in inches}

\item{width}{numeric, width in inches}

\item{dpi}{numeric, dpi of output}
}
\value{
ggplot bar graph
}
\description{
Generate overview of records (bar plot)
}
